microservices_backend/requirements.txt
langchain
langchain-core
langchain-groq
langgraph
langsmith
python-dotenv
langchain-community
fastapi
uvicorn
langchain-huggingface
sentence_transformers
pydantic
typing